Moved handlers in a route module
Renamed Config as MyConfig not to interfer with rocket's Config
Added some documentation
