+++ api/src/routes/mod.rs
@@ -1,0 +1,3 @@
+//! Contains every routes and handler
+
+pub mod user;
